Split is_legal into is_pseudo_legal and is_suicide. Renamed can_move is_trapped. Fixed castling again (kinda)
play_game now checks for (pseudo_legal && !is_suicide). The reason for this is that is_suicide can be called from is_trapped.

Castling also difficult to implement in any pretty way. is_suicide gets close, but it not close enough. is_suicide plays the move, "sees" it on the board, and decides what squares the king must have passed through to make the move (and checks those squares for check). If the king doesn't move, it just checks that square for check. This works the same no matter what the move is, castling or not, but it comes with one flaw: it does not check the home square for check, meaning you are allowed to castle out of check.