Changed is_hit() to is_hit_by(). Fixed castling (... again)
The king can no longer castle out of check. The reason I couldn't use is_hit() in K__() before was because that would make them infinitely corecursive. Now is_hit takes an integer bitmask which decides which pieces are used. For example, -1 uses all pieces (because every bit in the binary representation of -1 is set), while -33 is every piece except the king (because every bit except the 6th, the king's, is set.).

This does mean that the max number of piece types is 16 (there are 16 bits in an int)